refactor - move tests to existing module
