+++ t/unit/worker/test_state.py
@@ -1,3 +1,27 @@
+"""Unit tests for Celery worker state module.
+
+The worker state module (celery.worker.state) contains several global variables
+that are critical for tracking task state and worker operations:
+
+- reserved_requests: WeakSet of tasks that have been reserved but not yet executed
+- active_requests: WeakSet of tasks currently being executed
+- successful_requests: LimitedSet of task IDs that completed successfully
+- revoked: LimitedSet of task IDs that have been revoked
+
+These globals are fundamental to Celery's operation, affecting:
+- Task state management
+- Worker memory management (through WeakSets)
+- Task result tracking
+- Task revocation handling
+
+These tests ensure that:
+1. The globals maintain correct state throughout task lifecycle
+2. Memory is properly managed (no leaks in WeakSets)
+3. Set size limits are respected (LimitedSets)
+4. Task state transitions work correctly
+"""
+
+import gc
 import os
 import pickle
 import sys
@@ -14,6 +38,21 @@ from celery.utils.collections import LimitedSet
 from celery.worker import state
 
 
+class SimpleReq:
+    """Request class for testing.
+    
+    Simulates the minimal interface needed from celery.worker.request.Request
+    for testing worker state management.
+
+    Attributes:
+        id (str): Task ID used for tracking and set membership
+        name (str): Task name used by the worker for statistics
+    """
+    def __init__(self, name):
+        self.id = uuid()
+        self.name = name
+
+
 @pytest.fixture
 def reset_state():
     yield
@@ -21,6 +60,8 @@ def reset_state():
     state.revoked.clear()
     state.revoked_stamps.clear()
     state.total_count.clear()
+    state.reset_state()
+    gc.collect()
 
 
 class MockShelve(dict):
@@ -161,13 +202,6 @@ class test_Persistent:
             assert item in saved
 
 
-class SimpleReq:
-
-    def __init__(self, name):
-        self.id = uuid()
-        self.name = name
-
-
 @pytest.mark.usefixtures('reset_state')
 class test_state:
 
@@ -191,6 +225,132 @@ class test_state:
         for request in requests:
             state.task_ready(request)
         assert len(state.active_requests) == 0
+
+    def test_weakset_behavior(self):
+        """Test proper memory management of WeakSet globals.
+        
+        The reserved_requests and active_requests sets are WeakSets,
+        meaning they should automatically remove their entries when
+        the referenced objects are garbage collected.
+        
+        This test verifies that:
+        1. Items can be added to the WeakSet
+        2. Items are accessible while referenced
+        3. Items are automatically removed when dereferenced
+        """
+        # Create a reference and add to set
+        ref = SimpleReq("test-1")
+        state.reserved_requests.add(ref)
+        assert len(state.reserved_requests) == 1
+
+        # Verify item exists
+        items = list(state.reserved_requests)
+        assert len(items) == 1
+        assert items[0].name == "test-1"
+
+        # Remove reference and verify cleanup
+        items = None  # Remove reference from items list
+        ref = None  # Remove direct reference
+        gc.collect()  # Force garbage collection
+        assert len(state.reserved_requests) == 0
+
+    def test_limitedset_behavior(self):
+        """Test LimitedSet globals behavior and constraints.
+        
+        The successful_requests and revoked sets are LimitedSets,
+        which have a maximum size and can expire entries.
+        
+        This test verifies that:
+        1. Sets have correct type and configuration
+        2. Basic add/contains operations work
+        3. Size limits are enforced
+        """
+        # Test successful_requests
+        assert isinstance(state.successful_requests, LimitedSet)
+        assert state.successful_requests.maxlen == state.SUCCESSFUL_MAX
+        assert state.successful_requests.expires == state.SUCCESSFUL_EXPIRES
+
+        # Test basic add/contains
+        state.successful_requests.add("test-1")
+        assert "test-1" in state.successful_requests
+
+        # Test maxlen
+        for i in range(state.SUCCESSFUL_MAX + 10):
+            state.successful_requests.add(f"test-{i}")
+        assert len(state.successful_requests) <= state.SUCCESSFUL_MAX
+
+    def test_successful_requests_specific(self):
+        """Test specific behavior of successful_requests set.
+        
+        The successful_requests set tracks task IDs that have completed
+        successfully. This is used for duplicate task filtering and
+        result backend optimization.
+        
+        This test verifies that:
+        1. Task IDs are properly added
+        2. Membership testing works correctly
+        3. Non-existent tasks are properly reported as not present
+        """
+        # Test that it properly tracks successful task IDs
+        state.successful_requests.add("task-1")
+        state.successful_requests.add("task-2")
+        
+        assert "task-1" in state.successful_requests
+        assert "task-2" in state.successful_requests
+        assert "task-3" not in state.successful_requests
+
+    def test_revoked_specific(self):
+        """Test specific behavior of revoked set.
+        
+        The revoked set tracks task IDs that have been explicitly
+        revoked. This is critical for task cancellation and
+        duplicate task prevention.
+        
+        This test verifies that:
+        1. Task IDs are properly marked as revoked
+        2. Revocation status is correctly reported
+        3. Non-revoked tasks are properly reported as not revoked
+        """
+        # Test revocation tracking
+        state.revoked.add("task-1")
+        state.revoked.add("task-2")
+        
+        assert "task-1" in state.revoked
+        assert "task-2" in state.revoked
+        assert "task-3" not in state.revoked
+
+    def test_request_state_lifecycle(self):
+        """Test complete task request lifecycle through state transitions.
+        
+        Tasks move through several states during their lifecycle:
+        1. Reserved (pending execution)
+        2. Active (currently executing)
+        3. Complete (successful or failed)
+        
+        This test verifies that:
+        1. Tasks are properly tracked in each state
+        2. State transitions are handled correctly
+        3. Completed tasks are removed from tracking sets
+        4. Successful tasks are recorded in successful_requests
+        """
+        request = SimpleReq("lifecycle-1")
+        
+        # Test reservation
+        state.task_reserved(request)
+        assert request in state.reserved_requests
+        assert request.id in state.requests
+        
+        # Test acceptance
+        state.task_accepted(request)
+        assert request in state.active_requests
+        assert request.id in state.requests
+        
+        # Test completion (success)
+        state.task_ready(request, successful=True)
+        assert request.id in state.successful_requests
+        assert request not in state.active_requests
+        assert request not in state.reserved_requests
+        assert request.id not in state.requests
 
 
 class test_state_configuration():
